pyproject.toml
[tool.poetry]
name = "jobfinder"
version = "0.1.0"
description = ""
authors = ["Marius <m.gzegozevskis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
alembic = "^1.12.0"
sqlmodel="^0.0.8"
asyncpg="^0.28.0"
bcrypt="^4.0.1"
email-validator="^2.0.0.post2"
python-jose="^3.3.0"
python-multipart="^0.0.6"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
